Fix weird NodeJS 21 + undici behavior
It looks like something might override the `Response.prototype`
or attach different callbacks to the resonse that might actually
break if the original `value.apply(response, arguments)` is used
instead of the possibly monkey-patched runtime method.

As after all performance are the same but the code is smaller,
I think we're good with this change.